+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/settings/about/AboutModule.kt
@@ -3,12 +3,13 @@ package io.horizontalsystems.bankwallet.modules.settings.about
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import io.horizontalsystems.bankwallet.core.App
+import io.horizontalsystems.bankwallet.ui.helpers.TextHelper
 
 object AboutModule {
     class Factory : ViewModelProvider.Factory {
         @Suppress("UNCHECKED_CAST")
         override fun <T : ViewModel> create(modelClass: Class<T>): T {
-            return AboutViewModel(App.appConfigProvider, App.termsManager, App.systemInfoManager) as T
+            return AboutViewModel(App.appConfigProvider, TextHelper, App.termsManager, App.systemInfoManager) as T
         }
     }
 }
