+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/settings/about/AboutViewModel.kt
@@ -3,6 +3,7 @@ package io.horizontalsystems.bankwallet.modules.settings.about
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import io.horizontalsystems.bankwallet.core.IAppConfigProvider
+import io.horizontalsystems.bankwallet.core.IClipboardManager
 import io.horizontalsystems.bankwallet.core.ITermsManager
 import io.horizontalsystems.core.ISystemInfoManager
 import io.horizontalsystems.core.SingleLiveEvent
@@ -10,6 +11,7 @@ import io.reactivex.disposables.Disposable
 
 class AboutViewModel(
         private val appConfigProvider: IAppConfigProvider,
+        private val clipboardManager: IClipboardManager,
         termsManager: ITermsManager,
         systemInfoManager: ISystemInfoManager
 ) : ViewModel() {
@@ -17,8 +19,10 @@ class AboutViewModel(
     val openLinkLiveData = SingleLiveEvent<String>()
     val showShareAppLiveData = SingleLiveEvent<String>()
     val termsAcceptedData = MutableLiveData<Boolean>()
+    val showCopiedLiveEvent = SingleLiveEvent<Unit>()
 
     val appVersion = systemInfoManager.appVersion
+    val reportEmail = appConfigProvider.reportEmail
 
     var disposable: Disposable? = null
 
@@ -46,6 +50,11 @@ class AboutViewModel(
 
     fun onTellFriendsTap() {
         showShareAppLiveData.postValue(appConfigProvider.appWebPageLink)
+    }
+
+    fun didFailSendMail() {
+        clipboardManager.copyText(reportEmail)
+        showCopiedLiveEvent.call()
     }
 
 }
