--- app/src/main/java/io/horizontalsystems/bankwallet/modules/settings/contact/ContactPresenter.kt
@@ -1,28 +1,0 @@
-package io.horizontalsystems.bankwallet.modules.settings.contact
-
-import androidx.lifecycle.ViewModel
-
-class ContactPresenter(
-        val view: ContactModule.IView,
-        val router: ContactModule.IRouter,
-        private val interactor: ContactModule.IInteractor
-) : ViewModel(), ContactModule.IViewDelegate, ContactModule.IRouterDelegate {
-
-    override fun viewDidLoad() {
-        view.setEmail(interactor.email)
-        view.setWalletHelpTelegramGroup("@${interactor.walletHelpTelegramGroup}")
-    }
-
-    override fun didTapEmail() {
-        router.openSendMail(interactor.email)
-    }
-
-    override fun didTapWalletHelpTelegram() {
-        router.openTelegram(interactor.walletHelpTelegramGroup)
-    }
-
-    override fun didFailSendMail() {
-        interactor.copyToClipboard(interactor.email)
-        view.showCopied()
-    }
-}
