--- app/src/main/java/io/horizontalsystems/bankwallet/modules/settings/contact/ContactRouter.kt
@@ -1,17 +1,0 @@
-package io.horizontalsystems.bankwallet.modules.settings.contact
-
-import io.horizontalsystems.core.SingleLiveEvent
-
-class ContactRouter : ContactModule.IRouter {
-    val sendEmailLiveEvent = SingleLiveEvent<String>()
-    val openTelegramGroupEvent = SingleLiveEvent<String>()
-
-    override fun openSendMail(recipient: String) {
-        sendEmailLiveEvent.postValue(recipient)
-    }
-
-    override fun openTelegram(group: String) {
-        openTelegramGroupEvent.postValue(group)
-    }
-
-}
