--- app/src/main/res/layout/fragment_contact.xml
@@ -1,33 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:orientation="vertical"
-    android:background="?android:windowBackground"
-    android:clickable="true">
-
-    <androidx.appcompat.widget.Toolbar
-        android:id="@+id/toolbar"
-        android:layout_width="match_parent"
-        android:layout_height="?attr/actionBarSize"
-        app:navigationIcon="@drawable/ic_back"
-        app:title="@string/SettingsContact_Title" />
-
-    <io.horizontalsystems.views.SettingsView
-        android:id="@+id/mail"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="16dp"
-        app:icon="@drawable/ic_email"
-        app:title="@string/SettingsContact_Email" />
-
-    <io.horizontalsystems.views.SettingsView
-        android:id="@+id/walletHelpTelegramGroup"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        app:bottomBorder="true"
-        app:icon="@drawable/ic_telegram"
-        app:title="@string/SettingsContact_Telegram" />
-
-</LinearLayout>
